Data Prepper Event Language PoC
This creates a new ANTLR grammar for a Data Prepper Event Language. The current implementation attempts to build the Event Language on Data Prepper Expressions. However, the literal values that ANTLR generates are different, so these languages can't fully share code. However, this does work enough to set literal values to an event. Next, I'd like to find how to share the grammars more cleanly so that the code can work. Another problem is that Expressions expect that JSON pointers are evaluated immediately. This can't always work with an Event Language since we may perform put operations.

Signed-off-by: David Venable <dlv@amazon.com>